DevTools: Separate named hooks I/O and CPU work by thread
Keep CPU bound work (parsing source map and AST) in a worker so it doesn't block the main thread but move I/O bound code (fetching files) to the main thread so we can benefit from Network caching for source files.

Note that this change alone is not sufficient to be able to share the Network cache, but it will help making the subsequent work easier.
