--- packages/react-devtools-extensions/src/parseHookNames/parseHookNames.worker.js
@@ -1,11 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-import * as parseHookNamesModule from './parseHookNames';
-
-export const parseHookNames = parseHookNamesModule.parseHookNames;
-export const purgeCachedMetadata = parseHookNamesModule.purgeCachedMetadata;
