+++ packages/react-devtools-shared/src/constants.js
@@ -11,7 +11,7 @@
 export const __DEBUG__ = false;
 
 // Flip this flag to true to enable performance.mark() and performance.measure() timings.
-export const __PERFORMANCE_PROFILE__ = false;
+export const __PERFORMANCE_PROFILE__ = true;
 
 export const TREE_OPERATION_ADD = 1;
 export const TREE_OPERATION_REMOVE = 2;
